requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
asyncer==0.0.8
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
beartype==0.22.5
blinker==1.9.0
boto3==1.40.69
botocore==1.40.69
bracex==2.6
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
coverage==7.11.2
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.2.1
databricks-sdk==0.73.0
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
deepl==1.24.0
dill==0.4.0
discord.py==2.6.4
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
dockerfile-parse==2.0.1
docstring_parser==0.17.0
docutils==0.22.3
dspy==3.0.3
dspy-ai==3.0.3
e2b==2.6.3
e2b-code-interpreter==2.3.0
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.1
fastmcp==2.13.0.2
fastuuid==0.14.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gepa==0.0.7
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipdb==0.13.13
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
json_repair==0.53.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==25.6.0
kiwisolver==1.4.9
langdetect==1.0.9
linkify-it-py==2.0.3
litellm==1.79.3
loguru==0.7.3
lxml==6.0.2
magicattr==0.1.6
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.21.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mlflow==3.5.0rc0
mlflow-skinny==3.5.0rc0
mlflow-tracing==3.5.0rc0
more-itertools==10.8.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
omegaconf==2.3.0
openai==2.7.1
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.5.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
pyperclip==1.11.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
roma-dspy @ git+https://github.com/sentient-agi/ROMA.git@ea2886158df40943b1d5e28246a0ae850bc9e4d9
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
tenacity==9.1.2
textual==6.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vulture==2.14
watchfiles==1.1.1
wcmatch==10.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
